Cleanup cv (fixed flickering when loading caused by ssr)
cleanup isBrowser init

init projects page + cleanup

Fixes after review + update projects

remove dead code

hide projects (not ready) + readd doc

Removed project module
